모임 수정, 삭제 API 수정

+++ club/src/main/java/com/dbs/club/domain/meeting/MeetingService.java
@@ -3,6 +3,7 @@ package com.dbs.club.domain.meeting;
 
 import com.dbs.club.domain.common.exception.ErrorCode;
 import com.dbs.club.domain.meeting.exception.MeetingException;
+import com.dbs.club.domain.meetingjoin.MeetingJoinState;
 import com.dbs.club.domain.member.Member;
 import com.dbs.club.domain.member.MemberService;
 import com.dbs.club.infrastructure.meeting.MeetingRepository;
@@ -47,8 +48,12 @@ public class MeetingService {
     }
 
     @Transactional
-    public void updateMeeting(MeetingRequestDto.Update request, Long meetingId) {
+    public void updateMeeting(MeetingRequestDto.Update request, Long meetingId, Long memberId) {
         Meeting meeting = getMeeting(meetingId);
+
+        if (!meeting.isCreator(memberId)) {
+            throw new MeetingException(ErrorCode.MEETING_CAN_NOT_UPDATE);
+        }
 
         if (meeting.canNotUpdateMeetingDate()) {
            throw new MeetingException(ErrorCode.MEETING_CAN_NOT_UPDATE);
@@ -64,8 +69,14 @@ public class MeetingService {
     }
 
     @Transactional
-    public void deleteMeeting(Long meetingId) {
+    public void deleteMeeting(Long meetingId, Long memberId) {
         Meeting meeting = getMeeting(meetingId);
+
+        if (meeting.isNotCreator(memberId)) {
+            throw new MeetingException(ErrorCode.MEETING_CAN_NOT_DELETE);
+        }
+
+        meeting.getMeetingJoins().forEach(join -> join.updateStatus(MeetingJoinState.CANCEL));
 
         meeting.delete();
     }
